clean up code test
